converting destroy time var type to float rather than int.

+++ Assets/Scripts/CollectibleItem.cs
@@ -6,7 +6,7 @@ public class CollectibleItem : MonoBehaviour
 
 	public GUIManagement guiManagement;
 	public AudioSource pop;
-	public int destroyTime = 0.3f;
+	public float destroyTime = 0.3f;
 
 	public void Start ()
 	{
